Let the player move the stone if possible
Let the player move the stone if possible.

The positions the stone can move to is shown based on the "chosenStoneData" state inside the Board component.
This state gets set when the stone is being clicked, dragged or released inside the possible end position(s) for the stone.

The "endPositions" key value of the 'chosenStoneData" state gets used to render the PossibleMove component(s). This component shows stone(s) on a possible position where the chosen stone can move to.