Add starting functionality to move the stones
Add the first functionality to move the stones.

The functionality isn't working correctly yet, but it allows the stone to be moved based on the current player, and in bounds positions.

The functionality to capture opponents, isn't added yet.

The same is a bug that the player can move to a position where an enemy is located at, since the capture function isn't implemented yet.