Add this.gameInitialized to not run into issues
Add the this.gameInitialized, and set it to true when the componentDidMount() func gets called.

This is because the initializeStonePositions will be saved to the localstorage, and let the getStonesInformationData() func run, since the React strictmode will call the componentDidUnmount func, even when the component was not unmounted, which would cause issues to render the stones, since it will be an empty object that would be saved.