Fixed the func that  synchronize the settings
Improved the synchronization function that runs each time the app is opened by adding a check to determine if a key's value is a plain object. This ensures that recursion only occurs when both the local storage settings and the updated settings have an object as the value for a given key. Otherwise, the value is set to the initial default