Implement move tree calculation for checkers
Added multiple functions which allow us to calculate the move tree calculation for each stone.

For example getting the possible directions, and deleting mirrored directions for multi-capture moves.

We also added the functions to calculate the move tree, by also simulating the capture and revertion of the capture.

Also refactored comments and clarity of the code itself.